Put JS SDK related documentation under a new root, javascript-sdk/
