+++ src/app/website/community-and-support/community-and-support.view.html
@@ -56,7 +56,8 @@
 					<p><i class="big book icon"></i> <strong><a class="item" [routerLink]="[ '/documentation' ]">Documentation Home</a></strong> - Official home for the product documentation.</p>
 					<p><i class="big file text outline icon"></i> <strong><a href="quick-start-guide">Quick Start Guide</a></strong> - Learn how to download, configure, and run Carbon LDP and the
 						Workbench.</p>
-					<p><i class="big file text outline icon"></i> <strong><a class="item" [routerLink]="[ '/documentation/javascript-sdk-getting-started' ]">Getting Started with the JavaScript SDK</a></strong>
+					<p><i class="big file text outline icon"></i> <strong><a class="item" [routerLink]="[ '/documentation/javascript-sdk/getting-started' ]">Getting Started with the JavaScript
+						SDK</a></strong>
 						-
 						Learn how to use the
 						JavaScript SDK to build rich applications with pure client-side code.</p>
