+++ src/app/website/documentation/documentation.view.scss
@@ -84,7 +84,7 @@ documents {
 		color: gray;
 	}
 
-	sui-tabs /deep/ .ui.tabular.menu {
+	sui-tabs .ui.tabular.menu {
 		border-bottom: 0px;
 	}
