+++ src/app/website/documentation/home/home.view.html
@@ -143,15 +143,15 @@
 			</div>
 			<div class="ui bulleted list">
 				<div class="item">
-					<div class="header"><span [routerLink]="[ 'javascript-sdk-getting-started' ]">Getting started</span></div>
+					<div class="header"><span [routerLink]="[ 'javascript-sdk/getting-started' ]">Getting started</span></div>
 					<div class="summary">Guide to install Carbon JavaScriptSDK and start creating and manipulating data with its basic methods.</div>
 				</div>
 				<div class="item">
-					<div class="header"><span [routerLink]="[ 'javascript-sdk-object-model' ]">Object model</span></div>
+					<div class="header"><span [routerLink]="[ 'javascript-sdk/object-model' ]">Object model</span></div>
 					<div class="summary">An in depth description of the Carbon Object Model.</div>
 				</div>
 				<div class="item">
-					<div class="header"><span [routerLink]="[ 'javascript-sdk-contexts' ]">Contexts</span></div>
+					<div class="header"><span [routerLink]="[ 'javascript-sdk/contexts' ]">Contexts</span></div>
 					<div class="summary"> What is a context in Carbon JS SDK, how to declare,
 						access and modify it.
 					</div>
