+++ src/app/website/documentation/javascript-sdk.view.ts
@@ -1,0 +1,13 @@
+import { Component, ViewEncapsulation } from "@angular/core";
+
+@Component( {
+	selector: "javascript-sdk",
+	template: "<router-outlet></router-outlet>",
+	encapsulation: ViewEncapsulation.None,
+} )
+
+export class JavaScriptSDKView {
+
+}
+
+export default JavaScriptSDKView;
