+++ src/app/website/documentation/javascript-sdk/authorization.view.html
@@ -3,9 +3,9 @@
 		<div class="sixteen wide column">
 
 			<div class="ui breadcrumb">
-				<a class="section" [routerLink]="[ '..' ]">Documentation</a>
+				<a class="section" [routerLink]="[ '../..' ]">Documentation</a>
 				<i class="right chevron icon divider"></i>
-				<a class="section" [routerLink]="[ '../javascript-sdk' ]">JavaScript SDK</a>
+				<a class="section" [routerLink]="[ '..' ]">JavaScript SDK</a>
 				<i class="right arrow icon divider"></i>
 				<div class="active section">Authorization</div>
 			</div>
@@ -21,7 +21,7 @@
 					</header>
 
 					<div class="documentation-steps ui fluid steps">
-						<a class="step" [routerLink]="[ '../javascript-sdk-authentication' ]">
+						<a class="step" [routerLink]="[ '../authentication' ]">
 							<i class="arrow left icon"></i>
 							<div class="content">
 								<div class="title">Previous</div>
@@ -34,7 +34,7 @@
 								<div class="title">Authorization</div>
 							</div>
 						</div>
-						<a class="step" [routerLink]="[ '../javascript-sdk-access-points' ]">
+						<a class="step" [routerLink]="[ '../access-points' ]">
 							<i class="arrow right icon"></i>
 							<div class="content">
 								<div class="title">Next</div>
