+++ src/app/website/documentation/javascript-sdk/files.view.html
@@ -3,9 +3,9 @@
 		<div class="sixteen wide column">
 
 			<div class="ui breadcrumb">
-				<a class="section" [routerLink]="[ '..' ]">Documentation</a>
+				<a class="section" [routerLink]="[ '../..' ]">Documentation</a>
 				<i class="right chevron icon divider"></i>
-				<a class="section" [routerLink]="[ '../javascript-sdk' ]">JavaScript SDK</a>
+				<a class="section" [routerLink]="[ '..' ]">JavaScript SDK</a>
 				<i class="right arrow icon divider"></i>
 				<div class="active section">Uploading Files</div>
 			</div>
@@ -21,7 +21,7 @@
 					</header>
 
 					<div class="documentation-steps ui fluid steps">
-						<a class="step" [routerLink]="[ '../javascript-sdk-object-schema' ]">
+						<a class="step" [routerLink]="[ '../object-schema' ]">
 							<i class="arrow left icon"></i>
 							<div class="content">
 								<div class="title">Previous</div>
@@ -34,7 +34,7 @@
 								<div class="title">Files</div>
 							</div>
 						</div>
-						<a class="step" [routerLink]="[ '../javascript-sdk-querying' ]">
+						<a class="step" [routerLink]="[ '../querying' ]">
 							<i class="arrow right icon"></i>
 							<div class="content">
 								<div class="title">Next</div>
@@ -333,7 +333,7 @@
 						<p>TO BE DOCUMENTED</p>
 					</section>
 					<div class="documentation-steps ui fluid steps">
-						<a class="step" [routerLink]="[ '../javascript-sdk-object-schema' ]">
+						<a class="step" [routerLink]="[ '../object-schema' ]">
 							<i class="arrow left icon"></i>
 							<div class="content">
 								<div class="title">Previous</div>
@@ -346,7 +346,7 @@
 								<div class="title">Files</div>
 							</div>
 						</div>
-						<a class="step" [routerLink]="[ '../javascript-sdk-querying' ]">
+						<a class="step" [routerLink]="[ '../querying' ]">
 							<i class="arrow right icon"></i>
 							<div class="content">
 								<div class="title">Next</div>
