+++ src/app/website/documentation/javascript-sdk/javascript-sdk-home.view.html
@@ -1,56 +1,62 @@
-<div>
-	<div class="ui center aligned grid description">
-		<div class="ui container" id="javascript-sdk">
-
+<div class="ui main grid container" id="javascript-sdk">
+	<div class="row">
+		<div class="sixteen wide column">
 			<div class="ui breadcrumb">
 				<a class="section" [routerLink]="['..']">Documentation</a>
 				<i class="right chevron icon divider"></i>
 				<div class="active section">JavaScript SDK</div>
 			</div>
-
+		</div>
+	</div>
+	<div class="row">
+		<div class="sixteen wide column">
 			<div class="ui header">
 				<img class="ui small left floated image" src="assets/images/ico-javascript-150x150.gif">
 				<h2>JavaScript SDK</h2>
 			</div>
+		</div>
+	</div>
+	<div class="row">
+		<div class="sixteen wide column">
 			<div class="description">
 				<p>The JavaScript Software Developer's Kit, available from the npm package manager, allows you to manage RDF data using familiar JavaScript and
 					TypeScript programming techniques and tools. Build for execution within a web browser or Node.js.</p>
 			</div>
 			<div class="ui bulleted list">
 				<div class="item">
-					<div class="header"><span [routerLink]="[ '../javascript-sdk-getting-started' ]">Getting started</span></div>
+					<div class="header"><span [routerLink]="[ '../getting-started' ]">Getting started</span></div>
 					<div class="summary">Guide to install Carbon JavaScriptSDK and start creating and manipulating data with its basic methods.</div>
 				</div>
 				<div class="item">
-					<div class="header"><span [routerLink]="[ '../javascript-sdk-object-model' ]">Object model</span></div>
+					<div class="header"><span [routerLink]="[ '../object-model' ]">Object model</span></div>
 					<div class="summary">An in depth description of the Carbon Object Model.</div>
 				</div>
 				<div class="item">
-					<div class="header"><span [routerLink]="[ '../javascript-sdk-contexts' ]">Contexts</span></div>
-					<div clas="summary"> What is a context in Carbon JS SDK, how to declare, access and modify it.</div>
+					<div class="header"><span [routerLink]="[ '../contexts' ]">Contexts</span></div>
+					<div class="summary"> What is a context in Carbon JS SDK, how to declare, access and modify it.</div>
 				</div>
 				<div class="item">
-					<div class="header"><span [routerLink]="[ '../javascript-sdk-object-schema' ]">Object schema</span></div>
+					<div class="header"><span [routerLink]="[ '../object-schema' ]">Object schema</span></div>
 					<div class="summary">What is the object schema, how to define and use it.</div>
 				</div>
 				<div class="item">
-					<div class="header"><span [routerLink]="[ '../javascript-sdk-files' ]">Files</span></div>
+					<div class="header"><span [routerLink]="[ '../files' ]">Files</span></div>
 					<div class="summary">How to upload, download and describe files.</div>
 				</div>
 				<div class="item">
-					<div class="header"><span [routerLink]="[ '../javascript-sdk-querying' ]">Querying</span></div>
+					<div class="header"><span [routerLink]="[ '../querying' ]">Querying</span></div>
 					<div class="summary not-yet">Not yet documented</div>
 				</div>
 				<div class="item">
-					<div class="header"><span [routerLink]="[ '../javascript-sdk-authentication' ]">Authentication</span></div>
+					<div class="header"><span [routerLink]="[ '../authentication' ]">Authentication</span></div>
+					<div class="summary">Logging in, logging out and everything you need to know about it</div>
+				</div>
+				<div class="item">
+					<div class="header"><span [routerLink]="[ '../authorization' ]">Authorization</span></div>
 					<div class="summary not-yet">Not yet documented</div>
 				</div>
 				<div class="item">
-					<div class="header"><span [routerLink]="[ '../javascript-sdk-authorization' ]">Authorization</span></div>
-					<div class="summary not-yet">Not yet documented</div>
-				</div>
-				<div class="item">
-					<div class="header"><span [routerLink]="[ '../javascript-sdk-access-points' ]">Access Points</span></div>
+					<div class="header"><span [routerLink]="[ '../access-points' ]">Access Points</span></div>
 					<div class="summary not-yet">Not yet documented</div>
 				</div>
 			</div>
