+++ src/app/website/documentation/javascript-sdk/javascript-sdk-home.view.ts
@@ -1,0 +1,17 @@
+import { Component } from "@angular/core";
+
+import "semantic-ui/semantic";
+
+import template from "./javascript-sdk-home.view.html!";
+import style from "./javascript-sdk-home.view.css!text";
+
+@Component( {
+	selector: "javascript-sdk-home",
+	template: template,
+	styles: [ style ],
+} )
+
+export class JavaScriptSDKHomeView {
+}
+
+export default JavaScriptSDKHomeView;
