+++ src/app/website/documentation/javascript-sdk/object-model.view.html
@@ -2,9 +2,9 @@
 	<div class="row">
 		<div class="sixteen wide column">
 			<div class="ui breadcrumb">
-				<a class="section" [routerLink]="[ '..' ]">Documentation</a>
+				<a class="section" [routerLink]="[ '../..' ]">Documentation</a>
 				<i class="right chevron icon divider"></i>
-				<a class="section" [routerLink]="[ '../javascript-sdk' ]">JavaScript SDK</a>
+				<a class="section" [routerLink]="[ '..' ]">JavaScript SDK</a>
 				<i class="right arrow icon divider"></i>
 				<div class="active section">Object model</div>
 			</div>
@@ -19,7 +19,7 @@
 					</header>
 
 					<div class="documentation-steps ui fluid steps">
-						<a class="step" [routerLink]="[ '../javascript-sdk-getting-started' ]">
+						<a class="step" [routerLink]="[ '../getting-started' ]">
 							<i class="arrow left icon"></i>
 							<div class="content">
 								<div class="title">Previous</div>
@@ -31,7 +31,7 @@
 								<div class="title">Object Model</div>
 							</div>
 						</div>
-						<a class="step" [routerLink]="[ '../javascript-sdk-contexts' ]">
+						<a class="step" [routerLink]="[ '../contexts' ]">
 							<i class="arrow right icon"></i>
 							<div class="content">
 								<div class="title">Next</div>
@@ -226,7 +226,7 @@
 					</section>
 
 					<div class="documentation-steps ui fluid steps">
-						<a class="step" [routerLink]="[ '../javascript-sdk-getting-started' ]">
+						<a class="step" [routerLink]="[ '../getting-started' ]">
 							<i class="arrow left icon"></i>
 							<div class="content">
 								<div class="title">Previous</div>
@@ -238,7 +238,7 @@
 								<div class="title">Object Model</div>
 							</div>
 						</div>
-						<a class="step" [routerLink]="[ '../javascript-sdk-contexts' ]">
+						<a class="step" [routerLink]="[ '../contexts' ]">
 							<i class="arrow right icon"></i>
 							<div class="content">
 								<div class="title">Next</div>
