+++ src/app/website/documentation/quick-start-guide/quick-start-guide.view.html
@@ -277,7 +277,7 @@ This message shows that your installation appears to be working correctly.
 						<!-- TODO: Change this instructions to a more generic list of how to use the platform (through the REST API, for example) -->
 						<p>Of course, the app you created is not a functioning app; it's just the data space in Carbon for the application you now need to build. The next step is
 							to use the JavaScript SDK with your client-side code to build a client application. Head over to
-							<a class="item" [routerLink]="[ '/documentation/javascript-sdk-getting-started' ]">Getting Started with the JavaScript SDK</a> to learn how.</p>
+							<a class="item" [routerLink]="[ '/documentation/javascript-sdk/getting-started' ]">Getting Started with the JavaScript SDK</a> to learn how.</p>
 
 						<p>Note: If you are running your own local instance of the platform, you'll likely need to turn off https in the SDK. To do that configure the following setting
 							when creating your platform context (you'll understand what this means after reading the JavaScript SDK Getting Started guide):</p>
