+++ src/app/website/website.view.scss
@@ -15,6 +15,7 @@ website {
 
 		.main {
 			margin-top: 6em;
+			margin-bottom: 1em;
 		}
 
 	}
@@ -38,10 +39,6 @@ website {
 	img.logo {
 		margin-top: 2px;
 		margin-bottom: 2px;
-	}
-
-	article.ui.vertical.segment {
-		padding-bottom: 56px;
 	}
 
 	p.larger {
